+++ custom/text-enigma.php
@@ -3,3 +3,7 @@
 $text['point'] = 'zlobod';
 $text['hlaseniV'] = 'Zakázka';
 $text['hlaseniM'] = 'zakázka';
+
+
+$text['pridatvyjezd'] = 'nové hlášení z výjezdu';
+$text['pridatvyslech'] = 'nové hlášení z výslechu';