+++ inc/menu.php
@@ -9,7 +9,7 @@ function mainMenu()
   <ul class="'.$config['themeColor'].'">
 	  <li '.((searchTable(5)) ? ' class="unread"' : ((searchTable(5)) ? ' class="unread"' : '')).'><a href="/">Aktuality</a></li>
 	  <li '.((searchTable(6)) ? ' class="unread"' : ((searchTable(6)) ? ' class="unread"' : '')).'><a href="/board/">Nástěnka</a></li>
-      <li '.((searchTable(4)) ? ' class="unread"' : '').'><a href="reports.php">'.$text['hlaseniV'].' '.searchTable(4).'</a></li>
+      <li '.((searchTable(4)) ? ' class="unread"' : '').'><a href="/reports/">'.$text['hlaseniV'].' '.searchTable(4).'</a></li>
 	  <li '.((searchTable(1)) ? ' class="unread"' : ((searchTable(7)) ? ' class="unread"' : '')).'><a href="/persons/">Osoby '.searchTable(1).'</a></li>
 	  <li '.((searchTable(3)) ? ' class="unread"' : '').'><a href="/cases/">Případy '.searchTable(3).'</a></li>
 	  <li '.((searchTable(2)) ? ' class="unread"' : '').'><a href="groups/">Skupiny '.searchTable(2).'</a></li>